coalesce store, fix playcounts
